moved route top component
