+++ src/components/routes/RSVP.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import RouteTop from '../reusable/RouteTop';
+import RouteTop from './RouteTop';
 import RSVPForm from './rsvp/RSVPForm';
 
 class RSVP extends React.Component {
